+++ module_main/presentation_main/src/commonMain/kotlin/com/korchagin/presentation/models/PupilModel.kt
@@ -26,7 +26,7 @@ data class PupilModel(
         6 - преподаватель
     */
 
-    var subscription: Long,
+    var subscription: Int,
     /* ? = 0,
         5 - 1 месяц;
         2 - 3 месяца;
